pyproject.toml
[tool.poetry]
name = "pyyaml-typed"
version = "0.1.0"
description = "PyYAML dumper/loader using field descriptions in class comments"
authors = ["Anton Kachurin <katchuring@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/outcatcher/pyyaml-typed"
packages = [
    { include = "tyaml" },
]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.3.4"
coverage = "^5.0.3"
pylint = "^2.4.4"
pylint-quotes = "^0.2.1"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
